Moves checkbox-cell to case-tile-common
PR fixes
